location related schema renamings (WIP)
