+++ back/boxtribute_server/graph_ql/mutations.graphql
@@ -84,18 +84,4 @@ type Mutation {
   completeDistributionEventsTrackingGroup(
     id: ID!
   ): DistributionEventsTrackingGroup
-
-  # TODO: REMOVE THESE (REPLACED BY MORE SPECIFIC MUTATIONS)
-  # moveBoxToDistributionEvent(boxLabelIdentifier: ID!, distributionEventId: ID!): Box
-  # moveBoxFromDistributionEventToLocation(
-  #   boxLabelIdentifier: ID!
-  #   locationId: ID!
-  # ): DistributionEvent
-  # moveItemsFromDistributionEventToBox(
-  #   unboxedItemsCollectionId: ID!
-  #   boxLabelIdentifier: ID!
-  # ): DistributionEvent
-
-  # markDistributionEventAsComplete(distributionEventId: ID!): DistributionEvent
-  # END OF "REMOVE THESE (REPLACED BY MORE SPECIFIC MUTATIONS)"
 }
